Rewrite in C
There's really no reason to use C++ for such a small program. This is a complete rewrite that heavily optimizes the program. When I first wrote this, I didn't really know what I was doing with low-level code, so it allocated vast quantities of heap memory for things that needed a 243-byte static array and that kind of thing. It is now about 22,000 times faster (~50 hours to 8 seconds on my hardware) and does not allocate on the heap at all.